[tests:fix] Removed problematic import of tests in geo app
The geo app was importing test classes in the ready() method,
which caused some circular import issues in new development
branches we're working on.

This patch refactors internal testing logic to avoid that.
