+++ openwisp_controller/config/tests/test_admin.py
@@ -37,6 +37,7 @@ from .utils import (
     CreateConfigTemplateMixin,
     CreateDeviceGroupMixin,
     CreateDeviceMixin,
+    TestDeviceAdminMixin,
     TestVpnX509Mixin,
 )
 
@@ -53,47 +54,6 @@ User = get_user_model()
 Location = load_model("geo", "Location")
 DeviceLocation = load_model("geo", "DeviceLocation")
 Group = load_model("openwisp_users", "Group")
-
-
-class TestDeviceAdminMixin:
-    _device_params = {
-        "name": "test-device",
-        "hardware_id": "1234",
-        "mac_address": CreateConfigTemplateMixin.TEST_MAC_ADDRESS,
-        "key": CreateConfigTemplateMixin.TEST_KEY,
-        "model": "",
-        "os": "",
-        "notes": "",
-        "config-0-id": "",
-        "config-0-device": "",
-        "config-0-backend": "netjsonconfig.OpenWrt",
-        "config-0-templates": "",
-        "config-0-config": json.dumps({}),
-        "config-0-context": "",
-        "config-TOTAL_FORMS": 1,
-        "config-INITIAL_FORMS": 0,
-        "config-MIN_NUM_FORMS": 0,
-        "config-MAX_NUM_FORMS": 1,
-        # openwisp_controller.connection
-        "deviceconnection_set-TOTAL_FORMS": 0,
-        "deviceconnection_set-INITIAL_FORMS": 0,
-        "deviceconnection_set-MIN_NUM_FORMS": 0,
-        "deviceconnection_set-MAX_NUM_FORMS": 1000,
-        "command_set-TOTAL_FORMS": 0,
-        "command_set-INITIAL_FORMS": 0,
-        "command_set-MIN_NUM_FORMS": 0,
-        "command_set-MAX_NUM_FORMS": 1000,
-    }
-    # WARNING - WATCHOUT
-    # this class attribute is changed dynamically
-    # by other apps which add inlines to DeviceAdmin
-    _additional_params = {}
-
-    def _get_device_params(self, org):
-        p = self._device_params.copy()
-        p.update(self._additional_params)
-        p["organization"] = org.pk
-        return p
 
 
 class TestImportExportMixin:
@@ -293,34 +253,6 @@ class TestAdmin(
     object_model = Device
     object_location_model = DeviceLocation
     maxDiff = None
-    _device_params = {
-        "name": "test-device",
-        "hardware_id": "1234",
-        "mac_address": CreateConfigTemplateMixin.TEST_MAC_ADDRESS,
-        "key": CreateConfigTemplateMixin.TEST_KEY,
-        "model": "",
-        "os": "",
-        "notes": "",
-        "config-0-id": "",
-        "config-0-device": "",
-        "config-0-backend": "netjsonconfig.OpenWrt",
-        "config-0-templates": "",
-        "config-0-config": json.dumps({}),
-        "config-0-context": "",
-        "config-TOTAL_FORMS": 1,
-        "config-INITIAL_FORMS": 0,
-        "config-MIN_NUM_FORMS": 0,
-        "config-MAX_NUM_FORMS": 1,
-        # openwisp_controller.connection
-        "deviceconnection_set-TOTAL_FORMS": 0,
-        "deviceconnection_set-INITIAL_FORMS": 0,
-        "deviceconnection_set-MIN_NUM_FORMS": 0,
-        "deviceconnection_set-MAX_NUM_FORMS": 1000,
-        "command_set-TOTAL_FORMS": 0,
-        "command_set-INITIAL_FORMS": 0,
-        "command_set-MIN_NUM_FORMS": 0,
-        "command_set-MAX_NUM_FORMS": 1000,
-    }
 
     def setUp(self):
         self.client.force_login(self._get_admin())
