+++ openwisp_controller/connection/tests/test_admin.py
@@ -9,7 +9,6 @@ from swapper import load_model
 from openwisp_controller.connection.commands import ORGANIZATION_ENABLED_COMMANDS
 
 from ... import settings as module_settings
-from ...config.tests.test_admin import TestAdmin as TestConfigAdmin
 from ...tests import _get_updated_templates_settings
 from ...tests.utils import TestAdminMixin
 from ..connectors.ssh import Ssh
@@ -28,12 +27,6 @@ Group = load_model("openwisp_users", "Group")
 class TestConnectionAdmin(TestAdminMixin, CreateConnectionsMixin, TestCase):
     config_app_label = "config"
     app_label = "connection"
-    _device_params = TestConfigAdmin._device_params.copy()
-
-    def _get_device_params(self, org):
-        p = self._device_params.copy()
-        p["organization"] = org.pk
-        return p
 
     def _create_multitenancy_test_env(self):
         org1 = self._create_org(name="test1org")
@@ -287,6 +280,3 @@ class TestCommandInlines(TestAdminMixin, CreateConnectionsMixin, TestCase):
             response = self.client.get(url)
             self.assertContains(response, "https://example.com")
             self.assertNotContains(response, "owControllerApiHost = window.location")
-
-
-del TestConfigAdmin
